refacto: clean irrelevant implementation and file organization
Changelog:
- update pyproject to add 'asyncio'
- renamed:
  - generators.py -> pipeline.py
  - base/llm_class.py -> base/llm.py
  - base/text_generators.py -> base/text_generator.py
  - refacto organization of prompters/
    ├── answer
    │   ├── base.py
    │   └── with_retriever.py
    ├── eval
    │   └── fr.py
    ├── fact
    │   └── fr.py
    └── __init__.py

- removed:
  - config.toml: not yet needed
  - content of base/data_type.py moved to base/text_generator.py
  - irrelevant implementation in pipeline.py
  - deprecated prompter from version 0.033 [a1c1804da807539235addc1dd0ba1fcf95d1f7cf]

Recommendation:
the expe.py is to big and the name doesn't reveal the content. I think we should split it and move the data type in an other file.
