+++ src/ragtime/expe.py
@@ -457,6 +457,7 @@ class Expe(RagtimeList[QA]):
             logger.info(f'Expe saved as Spreadsheet to {path}')
             return path
     
+
 def analyse_expe_folder(path:Path):
     if not path.is_dir(): raise Exception(f'"{path}" is not a folder - please provide one')
     print(f'In "{path}":')
