+++ src/ragtime/prompters/fact/fr.py
