+++ src/ragtime/text_generators/eval_generator.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python3
 
-from ragtime.base.text_generators import ( TextGenerator )
-from ragtime.base.data_type import ( StartFrom )
+from ragtime.base.text_generator import (
+    TextGenerator,
+    StartFrom,
+)
+
 from ragtime.base.llm_class import ( LLM )
 
 from ragtime.expe import ( Eval, Facts, QA )
